requirements.txt
asgiref==3.7.2
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
absl-py==2.0.0
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
astunparse==1.6.3
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
databricks-cli==0.18.0
distro==1.8.0
docker==6.1.3
entrypoints==0.4
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.46.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.15.0
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-cloud-speech==2.23.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.1
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==6.11.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Mako==1.3.0
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
ml-dtypes==0.2.0
mlflow==2.8.1
numpy==1.26.2
oauthlib==3.2.2
openai==1.3.7
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
proto-plus==1.23.0
protobuf==4.23.4
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
querystring-parser==1.2.4
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlparse==0.4.4
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow-cpu==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
tiktoken==0.5.2
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
waitress==2.1.2
websocket-client==1.7.0
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
